Flag for execute
Add parameter flag_ to configure if you want to reachPreGrasp, reachAction or graspPlan without moving and only see the planing. Then if you want to execute that plan use executeAction.
Using move_group_->move() for what I have read in the API another plan it's made so you are not executing the plan you have seen.